Upgrade configcenter annotation
