+++ 3-extensions/configcenter/dubbo-samples-configcenter-annotation/dubbo-samples-configcenter-annotation-interface/src/main/java/org/apache/dubbo/samples/configcenter/annotation/api/AnnotationService.java
