+++ 3-extensions/configcenter/dubbo-samples-configcenter-annotation/dubbo-samples-configcenter-annotation-provider/src/main/java/org/apache/dubbo/samples/configcenter/annotation/ProviderApplication.java
@@ -1,0 +1,19 @@
+package org.apache.dubbo.samples.configcenter.annotation;
+
+import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@EnableDubbo
+public class ProviderApplication {
+
+    public static void main(String[] args) {
+        new EmbeddedZooKeeper(2181, false).start();
+        ZKTools.generateDubboProperties();
+        SpringApplication.run(ProviderApplication.class, args);
+
+        System.out.println("dubbo service started");
+    }
+
+}
