--- 3-extensions/configcenter/dubbo-samples-configcenter-annotation/src/main/resources/config-center.properties
@@ -1,32 +1,0 @@
-#   Licensed to the Apache Software Foundation (ASF) under one or more
-#   contributor license agreements.  See the NOTICE file distributed with
-#   this work for additional information regarding copyright ownership.
-#   The ASF licenses this file to You under the Apache License, Version 2.0
-#   (the "License"); you may not use this file except in compliance with
-#   the License.  You may obtain a copy of the License at
-#
-#       http://www.apache.org/licenses/LICENSE-2.0
-#
-#   Unless required by applicable law or agreed to in writing, software
-#   distributed under the License is distributed on an "AS IS" BASIS,
-#   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#   See the License for the specific language governing permissions and
-#   limitations under the License.
-
-# Add the following global external properties to `/dubbo/config/dubbo/dubbo.properties`
-dubbo.registry.address=zookeeper://127.0.0.1:2181
-#global config for consumer
-dubbo.consumer.loadbalance=random
-#global config for provider
-dubbo.protocol.port=20990
-dubbo.provider.timeout=5000
-dubbo.registry.simplified=true
-
-# Optional. Add the following application(Provider) specific external properties
-# to `/dubbo/config/configcenter-annotation-provider/dubbo.properties`
-dubbo.protocol.threadpool=fixed
-dubbo.protocol.threads=100
-
-# Optional. Add the following application(Consumer) specific external properties
-# to `/dubbo/config/configcenter-annotation-consumer/dubbo.properties`
-dubbo.consumer.timeout=6666
