--- 3-extensions/configcenter/dubbo-samples-configcenter-annotation/src/test/java/org/apache/dubbo/samples/configcenter/AnnotationServiceIT.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.dubbo.samples.configcenter;
-
-import org.apache.dubbo.samples.configcenter.annotation.AnnotationConsumer;
-import org.apache.dubbo.samples.configcenter.annotation.action.AnnotationAction;
-
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.test.context.ContextConfiguration;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
-
-@RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(classes = {AnnotationConsumer.ConsumerConfiguration.class})
-public class AnnotationServiceIT {
-    @Autowired
-    private AnnotationAction annotationAction;
-
-    @Test
-    public void test() throws Exception {
-        Assert.assertEquals("annotation: hello, world", annotationAction.doSayHello("world"));
-    }
-}
